[CST] Moving action types to single location (#22413)
* Moving action types into unified location
