+++ src/applications/claims-status/reducers/claim-sync.js
@@ -4,7 +4,7 @@ import {
   SET_CLAIM_DETAIL,
   SET_CLAIMS_UNAVAILABLE,
   SET_UNAUTHORIZED,
-} from '../actions';
+} from '../actions/types';
 
 const initialState = {
   synced: true,
