+++ src/applications/claims-status/reducers/notifications.js
@@ -5,7 +5,7 @@ import {
   SET_NOTIFICATION,
   CLEAR_ADDITIONAL_EVIDENCE_NOTIFICATION,
   SET_ADDITIONAL_EVIDENCE_NOTIFICATION,
-} from '../actions';
+} from '../actions/types';
 
 const initialState = {
   message: null,
